+++ test/test_helper.rb
@@ -68,13 +68,13 @@ class Minitest::Test
 
   def call_method(method, field, options)
     if ENV["ADAPTER"] == "enumerable"
-      Hash[@users.group_by_period(method, options) { |u| u.send(field) }.map { |k, v| [k, v.size] }]
+      Hash[@users.group_by_period(method, **options) { |u| u.send(field) }.map { |k, v| [k, v.size] }]
     elsif ENV["ADAPTER"] == "sqlite" && (method == :quarter || options[:time_zone] || options[:day_start] || options[:week_start] || Groupdate.week_start != :sun || (Time.zone && options[:time_zone] != false))
-      error = assert_raises(Groupdate::Error) { User.group_by_period(method, field, options).count }
+      error = assert_raises(Groupdate::Error) { User.group_by_period(method, field, **options).count }
       assert_includes error.message, "not supported for SQLite"
       skip
     else
-      User.group_by_period(method, field, options).count
+      User.group_by_period(method, field, **options).count
     end
   end