+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.2.1 (unreleased)
 
 - Fixed error with empty results in Ruby 2.7
+- Fixed deprecation warnings in Ruby 2.7
 
 ## 4.2.0 (2019-10-28)